#39 education modal component created and separated from student profilep page
